Moved all bluetooth related stuff to separate files, added app timer to check whether we need to signal the smartphone or not
